Implement New (game) functionality and user experience (with ability to change board dimensions)
Get rid of the "home" screen. It is not a good UE to have a separate screen for its current (and future) functionality. Remove ViewController.swift from the build (remains, for now, in the project), and make the Main storyboard instantiate the NumberTileGameViewController. Include "game controls" within the Main storyboard.

New (game) is available all of the time. This allows the player to start a new game both at the conclusion of the current one AND in the middle of a game that they wish to abandon.

Improve the GameModel design as part of adding this new functionality.