Share score functionality (Twitter, Facebook, and any other service configured on the user's system).
User can share a screen shot of the game board and score views (exclude the bottom toolbar), and they can do so anytime in the game (e.g. in-progress, game over). Employ the UIActivityViewController to gain a dynamic sharing ability, e.g. Facebook and Twitter, if configured in Settings, but also Mail, Save to camera roll, etc.